pyproject.toml
[tool.poetry]
name = "job-tracker"
version = "0.1.0"
description = "Automated internship and job tracker for Summer 2026"
authors = ["Yash Lad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg = {extras = ["binary"], version = "^3.1.13"}
alembic = "^1.12.1"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
apscheduler = "^3.10.4"
requests = "^2.31.0"
playwright = "^1.40.0"
selectolax = "^0.3.17"
beautifulsoup4 = "^4.12.2"
pyyaml = "^6.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
markdownify = "^0.11.6"
scikit-learn = "^1.3.2"
numpy = "^1.26.2"
sentry-sdk = "^1.38.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
vcrpy = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
